Rename modules/* -> api/*
