+++ src/api/stats/utils/endpoint-worker.js
