+++ src/api/stats/utils/flatten-result-set.js
