+++ src/api/stats/utils/flatten-result-set.spec.js
