+++ src/index.js
@@ -1,2 +1,5 @@
-import nba from './modules'
-export default nba
+import endpoints from './api'
+const { stats, data } = endpoints
+
+export { stats, data }
+export default endpoints
